requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
certifi==2023.5.7
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
Django==3.2.20
django-cors-headers==4.0.0
django-environ==0.10.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.3.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
multidict==6.0.4
phonenumbers==8.13.13
Pillow==9.5.0
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyotp==2.8.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
requirements-filter==0.0.0
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.4.4
stripe==5.4.0
twilio==8.2.2
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
yarl==1.9.2

